requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
charset-normalizer==2.0.12
flake8==4.0.1
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.2.0
mccabe==0.6.1
multidict==5.2.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.0.1
python-magic==0.4.25
typing_extensions==4.1.1
vt-py==0.13.1
yarl==1.7.2
zipp==3.6.0
